Fix for Versioneer
Versioneer requires that the package be built in a valid git
repository. `pip wheel .` copies the package to a temporary
directory then builds the wheel. The copying breaks the
repository state of the package because of the broken relative
path in the .git file of a submodule. To avoid this, we convert
the relative path to an absolute path.

Fixes #11